db/migrate_test.go
package db

import (
	"database/sql"
	"errors"
	"testing"

	"github.com/stretchr/testify/assert"
)

// MockDB is a mock implementation of the database/sql.DB interface for testing.
type MockDB struct {
	PrepareFunc func(query string) (*sql.Stmt, error)
}

func (m *MockDB) Begin() (*sql.Tx, error) {
	return nil, nil
}

func (m *MockDB) Exec(query string, args ...interface{}) (sql.Result, error) {
	return nil, nil
}

func (m *MockDB) Prepare(query string) (*sql.Stmt, error) {
	if m.PrepareFunc != nil {
		return m.PrepareFunc(query)
	}
	return nil, errors.New("Mock not implemented")
}

func (m *MockDB) Query(query string, args ...interface{}) (*sql.Rows, error) {
	return nil, nil
}

func (m *MockDB) QueryRow(query string, args ...interface{}) *sql.Row {
	return nil
}

func (m *MockDB) Close() error {
	return nil
}

func TestMigrateDB(t *testing.T) {
	// Create a mock DB for testing
	mockDB := &MockDB{}

	// Call the MigrateDB function with the mock DB
	MigrateDB(mockDB)

	// Add assertions or checks based on the behavior you want to test.
	// For example, you can check if certain queries were executed on the mock DB.
}

func TestProcessStruct(t *testing.T) {
	// Define a test model
	type dummytable struct {
		ID   int    `sql:"type:serial,primary key"`
		Name string `sql:"type:varchar(255)"`
	}

	// Call processStruct with the test model
	result := processStruct(dummytable{})

	// Add assertions or checks based on the expected behavior of processStruct.
	// For example, you can check if the result contains the expected table name and columns.
	assert.Equal(t, result.Name, "dummytables")
	assert.Equal(t, len(result.Columns), 2)
	assert.Equal(t, len(result.Values), 2)
	assert.Equal(t, result.Columns[0].Name, "ID")
	assert.Equal(t, result.Columns[0].Types, "serial primary key")
	assert.Equal(t, result.Columns[1].Name, "Name")
	assert.Equal(t, result.Columns[1].Types, "varchar(255)")
}

// Add more test cases as needed for other functions in your package.
